requirements.txt
click==8.1.3
docutils==0.19
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.2
numpy==1.24.2
Pillow==9.4.0
protobuf==4.21.12
Pygments==2.14.0
Werkzeug==2.2.3
